requirements.txt
asgiref==3.7.2
django==4.2.7
djangorestframework==3.14.0
drf-spectacular[sidecar]==0.26.5
gunicorn==21.2.0
packaging==23.2
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
black==23.11.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
python-environ==0.4.54
